backend/pyproject.toml
[tool.poetry]
name = "crypto-ninja"
version = "0.1.0"
description = ""
authors = ["Alexander Rykovskiy <alexander.rykovsky@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.1"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.30.6"
django-ninja = "^1.3.0"
pyjwt = "^2.9.0"
django-cors-headers = "^4.4.0"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
